+++ spec/spec_helper.rb
@@ -11,7 +11,6 @@ require 'active_record'
 ActiveRecord::Base.establish_connection('url' => 'sqlite3::memory:', 'pool' => 1)
 load 'spec/schema.rb'
 
-require File.expand_path(File.join(File.dirname(__FILE__), '../lib/formtastic/util'))
 require File.expand_path(File.join(File.dirname(__FILE__), '../lib/formtastic'))
 
 #rspec-rails 2-14 assumes a full rails install
@@ -469,14 +468,5 @@ RSpec.configure do |config|
 
   config.before(:example) do
     Formtastic::Localizer.cache.clear!    
-  end
-
-  config.before(:example) do
-    allow(Formtastic.deprecation).to receive(:deprecation_warning).and_call_original
-    # TODO: Remove this in Formtastic 4
-    [ :action_class, :standard_action_class_name, :custom_action_class_name,
-      :input_class, :standard_input_class_name, :custom_input_class_name ].each do |method|
-      allow(Formtastic.deprecation).to receive(:deprecation_warning).with(method, instance_of(String), instance_of(Array))
-    end
   end
 end